refactor: extract usd t configuration provider
